title scene(setting window - volume controller)
