+++ GMDgame/Assets/Script/TitleScripts/MusicManager.cs
@@ -1,0 +1,41 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.Audio;
+
+public class MusicManager : MonoBehaviour
+{
+    private AudioSource audioSource;
+    private GameObject[] musics;
+    public AudioMixer mixer;
+
+    //guarantee music object uniqueness
+    //this object will not be destroyed after scene change
+    private void Awake()
+    {
+        musics = GameObject.FindGameObjectsWithTag("Music");
+
+        if (musics.Length >= 2)
+            Destroy(this.gameObject);
+
+        DontDestroyOnLoad(transform.gameObject);
+        audioSource = GetComponent<AudioSource>();
+    }
+
+    //private void Start()
+    //{
+    //    mixer.SetFloat("MusicVolume", Mathf.Log10(0.3f) * 20);
+    //}
+
+    public void PlayMusic()
+    {
+        if (audioSource.isPlaying) return;
+        audioSource.Play();
+    }
+
+    public void StopMusic()
+    {
+        audioSource.Stop();
+    }
+}
