+++ GMDgame/Assets/Script/TitleScripts/MusicManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e0474677011d5f743a7048304547fede
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
