+++ GMDgame/Assets/Script/TitleScripts/SetVolume.cs
@@ -1,0 +1,26 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Audio;
+using UnityEngine.UI;
+
+public class SetVolume : MonoBehaviour
+{
+    public AudioMixer mixer;
+    public Slider slider;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        slider.value = PlayerPrefs.GetFloat("MusicVolume", 0.75f);
+    }
+
+    public void SetLevel(float sliderValue)
+    {
+        //float sound 
+        mixer.SetFloat("MusicVolume", Mathf.Log10(sliderValue) * 20);
+        PlayerPrefs.SetFloat("MusicVolume", sliderValue);
+    }
+
+    
+}
