+++ GMDgame/Assets/Script/TitleScripts/SetVolume.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 856870508c9f66f44acecdd41d414930
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
