+++ GMDgame/Assets/Script/TitleScripts/SettingManager.cs
@@ -2,7 +2,6 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
-using UnityEngine.Audio;
 
 public class SettingManager : MonoBehaviour
 {
@@ -11,24 +10,6 @@ public class SettingManager : MonoBehaviour
 
     [SerializeField]
     private GameObject settingWindow;
-
-    [SerializeField]
-    //private AudioMixer mixer;
-    private AudioSource audioSource;
-    private GameObject[] musics;
-
-
-    private void Awake()
-    {
-        musics = GameObject.FindGameObjectsWithTag("Music");
-
-        if (musics.Length >= 2)
-            Destroy(this.gameObject);
-
-        DontDestroyOnLoad(transform.gameObject);
-        audioSource = GetComponent<AudioSource>();
-    }
-
 
     // Start is called before the first frame update
     void Start()
@@ -43,14 +24,5 @@ public class SettingManager : MonoBehaviour
     }
 
 
-    public void PlayMusic()
-    {
-        if (audioSource.isPlaying) return;
-        audioSource.Play();
-    }
-
-    public void StopMusic()
-    {
-        audioSource.Stop();
-    }
+    
 }
